Add handling for firewalld's StrictForwardPorts setting
This mostly amounts to throwing a sensible error in cases where
the setting is enabled and the user has requested ports be
forwarded.

There are no tests at present because firewalld 2.3.x is
currently restricted to Rawhide, and is required for effective
testing.

A manpage has also been added with details on how to work with
the StrictForwardPorts setting (as well as some other tidbits on
firewalld + Podman integration).

Signed-off-by: Matt Heon <mheon@redhat.com>